Adds `content_finder` method to hide away Nokogiri
The implementation of this Gem shouldn't require as much knowledge of how Nokogiri parsing works. The `get_listing` method is deprecated in favor of `content_finder` (same name as the online utility on the SODA XML Team console). This new method returns an array of document objects that are easier to work with.